+++ src/pixel.c
@@ -3,5 +3,5 @@
 int PixelGet_(int x, int y)
 { return PixelGet_internal_(x,y); }
 
-void PixelSet_(int x, int y, int color)
+void PixelSet_(int x, int y, uint8_t color)
 { PixelSet_internal_(x,y, color); }
